Remove Terraform reference
It's not possible to spin up an AWS Marketplace AMI from Terraform due to the need for the user to have previously accepted the terms of the AMI. Since most users will not have done this already the Terraform process will fail. Remove this reference and consider an alternative approach for introducing automation with Terraform.